Moving font size and styles to separate styles directory, adding word wrap
