+++ src/components/Styles/adaptiveFontSize.tsx
